+++ native/src/main/java/io/ballerina/stdlib/sql/utils/RecordIteratorUtils.java
@@ -59,15 +59,22 @@ public class RecordIteratorUtils {
                         .getNativeData(Constants.COLUMN_DEFINITIONS_DATA_FIELD);
                 Utils.updateBallerinaRecordFields(resultParameterProcessor, resultSet, bStruct, columnDefinitions);
                 return bStruct;
-            } else {
-                return null;
             }
+            // Stream has reached the end, we clean up the resources, here any error from closing the stream is ignored.
+            closeResult(recordIterator);
+            return null;
         } catch (SQLException e) {
+            // Stream throws an error, we clean up the resources, here any error from closing the stream is ignored.
+            closeResult(recordIterator);
             return ErrorGenerator.getSQLDatabaseError(e, "Error when iterating the SQL result");
         } catch (ApplicationError e) {
+            // Stream throws an error, we clean up the resources, here any error from closing the stream is ignored.
+            closeResult(recordIterator);
             return ErrorGenerator.getSQLApplicationError("Error when iterating the SQL result. "
                     + e.getMessage());
         } catch (Throwable throwable) {
+            // Stream throws an error, we clean up the resources, here any error from closing the stream is ignored.
+            closeResult(recordIterator);
             return ErrorGenerator.getSQLApplicationError("Error when iterating through the " +
                     "SQL result. " + throwable.getMessage());
         }